pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-telegram-bot"
version = "0.1.0"
description = "MCP server that exposes a Telegram bot"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Dario Clavijo", email = "clavijodario@gmail.com"}
]
dependencies = ["fastmcp", "python-telegram-bot>=20.0"]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "hatch",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio",
    "hypothesis",
]
lint = [
    "ruff",
    "mypy",
]
all = ["mcp-telegram-bot[dev,test,lint]"]

[project.scripts]
mcp-telegram-bot = "mcp_telegram_bot.__main__:main"

[project.urls]
Homepage = "https://github.com/daedalus/mcp-telegram-bot"
Repository = "https://github.com/daedalus/mcp-telegram-bot"
Issues = "https://github.com/daedalus/mcp-telegram-bot/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_telegram_bot"]

[tool.hatch.build.targets.sdist]
include = ["src/mcp_telegram_bot"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "TCH", "N", "C4", "ARG"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG001", "ANN001", "ANN201"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-fail-under=80"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude = ["tests/*", "*/__init__.py"]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
